Updated openstack/openstack Project: openstack/keystone  b3645a74219d332658093d49d9c5427c47a9d065
Add audit ids to tokens

Add audit identifiers to each token. If the token is a rescoped token
the audit identifier will be a list where index 0 being the token's 
audit id and index 1 is the initial token in the chain's audit_id.

If the token is not rescoped, the token's audit_ids will only contain
the audit_id of the token itself (at index 0).

This change is to resolve a method of performing token revocations
on individual tokens without needing to rely on the low
resolution of token expires time.

DocImpact

Change-Id: If5fdf187bd525f273281f9fdf53bdc8865164289
bp: non-persistent-tokens