Updated openstack/openstack Project: openstack/nova  8d4ee0af1acf5e6a9ac0582ff9caa823640f5f54
Remove _instance_update usage in _build_instance

The _build_instance method has an instance object so
just call instance.save() rather than _instance_update
which does a separate call to conductor to update the
instance updates the resource tracker, which is
unnecessary in this case. Compare to the _build_resources
method.

Part of blueprint compute-manager-objects-juno

Change-Id: Ic350a424f83638e8129d55dda168913798ebc31a