Updated openstack/openstack Project: openstack/heat  fc118ae8cbea95a674bb59fb6aee2c5d83eb2b69
Convert CFN waitcondition API to resource_signal

Convert the CFN waitcondition API to use the resource_signal API,
which will enable removal of the duplicate metadata_update paths
in the engine.  A side-effect of this is that we're using the
signal interfaces which create more descriptive events when
handling the signals, ref bug #1340894.

Change-Id: I1cf7bac78846a4292822bd26d01a7a148a9a0a0d
Partial-Bug: #1340894