Updated openstack/openstack Project: openstack/horizon  cc945f907081d52f861d8f89c1a8c70c43bc8f85
Fix odd row last cell border colors on Firefox

Firefox was allowing the background color to overlap the borders which
made it look like there was not a border on the cell. This shows up
when the rows are striped, or on a row hover. Fixed by explicitly
adding background-clip:padding-box, which should have been the
default behavior.

Change-Id: Ifd64cbea09ba5288da652b5306263766b5953358
Closes-bug: 1353695