Updated openstack/openstack Project: openstack/python-troveclient  33c76fab16b80f32a04f83e1653d4f809c22d3bd
Add a drop config-grp command in update instance

As per the current code in the client, there is no option to detach
the configuration group of the instance. This change gives a user the
option to explicitly drop the config group of the instance as part of
the instance update call.

Change-Id: I7defd110549df87afd267d3ebe998904ff6b7483
Closes-Bug: 1359359