Updated openstack/openstack Project: openstack/horizon  a9fcc934b8cc625db9b40b0d0ba708149c586a98
Fixed issue of 'terminate instance' with fresh login

When fresh login horizon, go to Instances page in project or
admin panel, if try to terminate instance, it does
not do anything. The symptom also applies to all the
tables.BatchAction in instance page, for example,
Hard Reboot, Soft Reboot, Suspend and etc.

Similiar issue for resizing instance, once issue resize
request, click on the confirm button, it doesn't do anything for
fresh login as well.

When trigger one of the actions on the instance page
the very first time after login, the filter_string_session
is None during the process of checking the filter info.
The original condition checking causes filter change is True so it
thinks it is the filterAction, so it gets redirected to the instance
page instead of handling the action request.

Fixed in get_server_filter_info to check
filter_string_session is not None

Closes-Bug: #1356517
Closes-Bug: #1356546

Change-Id: I2c093b70c781a63631bd6513b83feb2a698392a6