Updated openstack/openstack Project: openstack/nova  c594340359e3d148b6c349c388f10327ac7b8bfa
Log warning message if volume quota is exceeded

It would be useful to the user if a warning message is logged
when the volume quota is exceeded instead of a debug message.
This would allow the root cause to be more easily determined
instead of turning on the debug option.

Change-Id: I19dc078cbacf8f39480e8a7f4c4edb764e4696eb
Closes-Bug: #1349268