Updated openstack/openstack Project: openstack-dev/devstack  50d4158d6cb67e052131698f3526dbabcd2cc3dc
Use tox for bashate testing

Create a tox.ini file to install bashate into a virtualenv for running
tests. It can be run with a command such as "tox -v -ebashate". The find
command being used is equivalent to the default list of files discovered
by bashate plus the exclusion of the .tox/ directory. Also add the .tox/
directory to the .gitignore file.

Change-Id: I94cab4e4a9118d62edfce29ead1f1e2a5f48e529