Updated openstack/openstack Project: openstack/oslo.db  4c7fc3e685e92880f48259fbec9ed72b471af89e
Remove reliance on create_engine() from TestsExceptionFilter

As we are continuing to add functionality to create_engine()
which in some cases means that backend-specific SQL is
emitted on connect, the practice of repurposing a
SQLite or other engine to act like another one for the
purposes of a test is becoming less feasible.   Here,
TestsExceptionFilter is altered to not rely upon the full
session.create_engine() process for unit tests, instead
calling upon just those listeners being tested.

Change-Id: I0ece42f956c59eab7a6cb86419d8d359250d5e71