Updated openstack/openstack Project: openstack/nova  92d287e4452fb5a51af1d071efac06a2cf735967
neutronv2: treat instance as object in deallocate_for_instance

The deallocate_for_instance method is given an instance object
but the neutronv2 API was treating it like a dict. This changes
the method to use dot notatoin for accessing fields on the
object and updates the unit tests to pass an instance object
rather than a dict.

Part of blueprint compute-manager-objects-juno

Change-Id: I784d1d59c23da8485d355251aba26b98bd9c7a16