Updated openstack/openstack Project: openstack/oslo-incubator  83c4098015d907a0d35e440f736bcc804f35c8a6
Mask passwords in exceptions and error messages

When a ProcessExecutionError is thrown by processutils.execute(), the
exception may contain information such as password. Upstream
applications that just log the message (as several appear to do) could
inadvertently expose these passwords to a user with read access to the
log files. It is therefore considered prudent to invoke
strutils.mask_password() on the command, stdout and stderr in the
exception. A test case has been added to ensure that all three are
properly masked.

OSSA is aware of this change request.

Submitted to oslo.concurrency in
Ie122db5f19802f519b96ed024ab3f2b5eede3eee

Change-Id: I173dfb865e84eb7dee54a22c76db1e4f125a0a8a
Closes-Bug: #1343604