Updated openstack/openstack Project: openstack/oslo.messaging  48a9ba4e27bcf237f8b9032e70d79a3e22a429aa
Fix Python 3 testing

Raising SkipTest at import time fails in recent version of the test
suite. We change that by being more picky at one should or should not be
tested based on what we managed to import; the upside is that we test a
few more things under Python 3 now.

Change-Id: I8971c7291cb261faf00fec73cf53b6132bdaf948