Updated openstack/openstack Project: openstack/trove  eb8cdbcf4595fcc17efa022c24cf9bad17946146
Move usage_timeout out of guest options

In an effort to further separate the guest and task manager,
the usage_timeout has been moved into the common cfg. The
datastores have no need for this. the default is now set as
the largest timeout, cassandra.

DocImpact: This change modifies a configuration parameter.
partially implements: blueprint refactoring-datastore-options-in-cfg
Co-Author: Michael Basnight <mbasnight@gmail.com>

Change-Id: I3b4b2a9bffb40df4531a78c29d1e155e59436948