Updated openstack/openstack Project: openstack/tempest  0ee7e4ca2fb7451b80b7a15fc55275194eb06b58
Add cases for Swift scenario test

Based on current implement, we have already implemented the
basic ops of swift. However, as a object storage service,
object download and verification are key function of it. So
this patch is aim to enhance that to get a complete end to end
function flow.

Change-Id: Id0bdcd2105ad90ea1500c7cbd1acdbe19e5aa361