Updated openstack/openstack Project: openstack/nova  cae732cf5bcf7c12591c4ccc74353b0fc58eb2c5
objects: Add base test for obj_make_compatible()

A previous commit fixed a bug in obj_make_compatible() of the
InstanceGroup object.  While fixing that, I added a test targeted at
that object.  This patch adds a test that calls obj_make_compatible()
on every past version of every object.  It doesn't actually test that
the data is converted, but it makes sure that calling the method
doesn't blow up on something basic like expecting the wrong format of
the version parameter.

Only one object had to change for this test to work, and that was a
test object.  It assumed a field was always set, and it's not in the
case of this test.  However, the field in question is not actually set
as nullable, so it *should* be set.  However, the way these methods
are implemented elsewhere, they never assume the field to be set and
present in the primitive version of the object.

Change-Id: Id56b2f9763cfcb99a3c4a2d56329485c5e9c3b5c