Updated openstack/openstack Project: openstack/nova  d65baa21e3382158cb7b4f9db4c7f89810b56f43
Remove duplicated code in test_user_data

In test_user_data, there is duplicated code for creating a request.
In addition, there is an internal method which works as almost same
as the duplicated code.
This patch makes the tests use the internal method and removes the
duplicated code for cleanup.

Change-Id: Ie87cc14a3ac67bc1262de155dc0a1dcf5c5e7d9a