Updated openstack/openstack Project: openstack/keystone  05c47c807be9fb1970cce84000992d3ac53cf37f
Back off initial migration to 35

Upgrading from the Havana release to Juno is failing because the
initial Juno migration doesn't include a couple of migrations that
weren't in the initial Havana release but were added afterwards.

This change adds the 036 migration from Havana back in and moves
the initial Juno migration back to 035.

Change-Id: I1045696d66d7a63587aaaf70230a5eaf46e7d65d
Partial-Bug: #1357498