Updated openstack/openstack Project: openstack/heat  7d3ffe21ded67e538e23de6eee61ba947d3b433c
Provide access to outputs from OS::Heat::AutoScalingGroup

Allow get_attr on OS::Heat::AutoScalingGroup to return a list of
corresponding attributes from each of the base resources. For example,

  {'get_attr': ['my_asg', 'outputs_list', 'networks', '<network_name>', 0]}

will return a list of the first IP address assigned to each server in the
group. This is modelled on similar functionality in the
OS::Heat::ResourceGroup resource. Similarly:

  {'get_attr': ['my_asg', 'outputs', 'networks', '<network_name>', 0]}

returns a mapping from resource names to the corresponding IP addresses.

Change-Id: Ib9e830e0c54602ac456d64179e1df30f0778906d
Closes-bug: #1356544