Updated openstack/openstack Project: openstack/horizon  e56baa97131e3139bd7e3400104b3152c5d98273
Fix handling of missing external network

When an external network can't be found, project/routers/index.html will
result in a type error caused by calling exceptions.handle when no
exception is present.  In this case, an error message should be
displayed, and the missing-network should be indicated to the user.

Also added tests for this case.

Change-Id: I3aaa64de28738051ad84cc587fbb6cfd546f4662
Closes-Bug: 1220345