Updated openstack/openstack Project: openstack/nova  95e7e2f33b08d9cddd3dd8c22ca18c16ac8b698f
VMware: refactor spawn() code to build a new VM

Move the spawn() code which builds a new VM into
vmops.build_virtual_machine().

We remove the duplicate check of instance_name. Other than that this
is straight code motion.

In addition, we add a basic test of the function of
build_virtual_machine, which tests that all given configuration is
reflected in the resulting virtual machine.

This change has been split out of
https://review.openstack.org/#/c/87002/, which was written by Shawn
Hartsock.

partial blueprint vmware-spawn-refactor

Co-authored-by: Shawn Hartsock <hartsocks@acm.org>
Change-Id: Id9c95966eefb644fdaf36465e7b3d3583826314a