Updated openstack/openstack Project: openstack/horizon  0ee604c179f527774cf30fda05013a6ce7a3fe3e
Add dropdown actions to detail page

Add support for performing actions on an object from the object's
detail page. Actions are displayed in the page header, in a
dropdown menu. As a POC, this patch enables dropdown actions on a
detail page of a volume (/project/volumes/volume-id/).

Partially implements: blueprint detail-pages-ia

Change-Id: Ie21c63d86c2806d9689ae7f1a67b67c83710c108