Updated openstack/openstack Project: openstack/nova  ebeca96fcececdd966c3a3fd5acdf5ec55895d47
Avoid refreshing PCI devices on instance.save()

Currently the way we refresh Instance.pci_devices causes current-version
PciDevice and PciDeviceList objects to be sent back to a potentially old
client. This is because remotable methods don't undergo backporting
obligatorily in the way that remotable_classmethod ones do. This is
currently a problem for pci_devices and security_groups, although it is
handled differently for things like info_cache so that it's not a
problem.

The resolution to this is rather complicated, so this patch just avoids
the implicit update (which is probably never used and actually generating
some more DB traffic) for the time being.

Change-Id: I33e584c42c9484912ccc3d38b67a6a7ee3d69a10