Updated openstack/openstack Project: openstack/neutron  2f7932055247df3ac5ec4783e029cfeda0466f71
Change autogenerate to be unconditional

Template for generating db migration scripts are now not run according
to used plugins. Target environment is set to use all available models
but also used plugins are kept for backward compatibility.

Part of this patch is script that will drop unused tables to avoid
generating the drop in future migration scripts.

Change-Id: I90a7d3416de24f2317ccdeb828bb8bd973ee5e61
Closes-bug: #1346638