Updated openstack/openstack Project: openstack/neutron  1a16ba4ecd717d4c4c9bbe8c6a61abba202b125a
VPNaaS: Enable UT cases with newer oslo.messaging

Now that 1.4.0.0a3 of oslo.messaging has been added to Neutron
requirements, the unit test cases in VPNaaS code can be
uncommented and included in the tests. These tests ensure that
the correct validator is called, when validation is performed.
This adds similar tests to the reference and Cisco VPNaaS
implementation.

Change-Id: I62e3f8f3ec5fcceccd13891b1aa142869e1d88a1
Closes-Bug: 1349829