Updated openstack/openstack Project: openstack/keystone  e971c89dc5d736a21decb26efe4a2b692ba5bf4d
Back off initial migration to 34

Upgrading from the Havana release to Juno was failing because the
initial Juno migration didn't include a couple of migrations that
weren't in the initial Havana release but were added afterwards.

This change adds the 035 migration from Havana back in and moves
the initial Juno migration back to 034.

Change-Id: I012b3f707659557d1774bae78b1dd652bb1a4dd7
Closes-Bug: #1357498