Updated openstack/openstack Project: openstack/nova  993a7678f186c0a43af41a78cbb04df3e2db0cf4
Network: add in a new network type - DVS

The support enables DVS network backends. In this case neutron
will return a vif type of DVS.

The name of the VC port group will be of the following format:
    <network-name>-<network-uuid>

The network name is so that the prefix will be easy for people to
understand. The UUID is to ensure that it will be unique.

Change-Id: Id125797632591bd799e215da7cbae07154e90344