Updated openstack/openstack Project: openstack/python-keystoneclient  b939dd9922abdb5257f785fca4e1d41741ddb078
Make tests run against original client and sessions

The managers have a bad habit of reaching back into the client and
making assumptions about the values that are saved there. These
assumptions are not always correct when we use the session object.

Test all the versioned managers against a client that was constructed
with the old method and with a keystoneclient session object and a
keystoneauth1 session object.

Change-Id: I93a26db7ae7e4d887aa815108be71c72b4a1f2bb