Updated openstack/openstack Project: openstack/nova  c3ad01b65fd3044cb4240c848afbfbc36222f8f5
Fix NotImplementedError in floating-ip-list

A list of existing floating IP (nova floating-ip-list) fails
when using VMwareVCDriver/LibvirtDriver and nova-network.
The following error is thrown: "NotImplementedError: Cannot load
'instance' in the base class". The problem is caused by
_normalize_ip, which tries to set an 'instance' field in
FloatingIP that is needed by _translate_floating_ip_view. The
instance is derived from the FixedIP that is assigned to the
'fixed_ip' field in FloatingIP. However, FixedIP may not have
an 'instance' field set, which can cause the NotImplementedError.

The following patch removes _normalize_ip and assigns the
'instance_id' required by _translate_floating_ip_view to
the 'instance_uuid' in FixedIP, since only the instance ID
is required and not the entire instance object.

Change-Id: I2d7937e969bd87e53557842a22f3861f0d307b5e
Closes-Bug: #1356051