Updated openstack/openstack Project: openstack/nova  46ce77bb84bca7c109e1869f785e78fd437950e5
Remove duplicated negative factors in keypair test

In test_keypair_create_fail_policy, there are two negative factors.
One is a policy violation, and another is a bad request body.
The purpose of the test is checking the policy violation, so a bad
request body is not right for this test.
This patch removes a bad request body from the test.

Change-Id: Ia165e08410b0576950efe83e5ab6a92c9c7f6a50