Updated openstack/openstack Project: openstack/nova  c48b7b9cb79266a87b21a0528eee49d03bf1980c
Preserve exception text during schedule retries

This fix preserves the exception text (as intended) when the number
of scheduling retries exceeds the maximum attempts.  We don't pop
the exception message anymore in _log_compute_error so that the
final "max attempts" error message has an opportunity to dump it
as well.

This defect caused the exception text (from the compute process) to
ALWAYS be None, thus making it difficult to triage compute errors.

Change-Id: I5c7cac1d79dc583cca523270bb0a96d3353ca1f6
Closes-Bug: #1361097