Updated openstack/openstack Project: openstack/heat  b9b9c4864a25893571c0e5ca9893abceb809e421
boolean parameter type is not recognized in the API

The boolean type that was added to Heat templates is not recognized
in the API layer. Because of this, the API passes the schema type
(which is lowercase 'boolean') directly to the client, while all
other recognized types are passed with camel case.

Change-Id: I0d31401456c0f5cd0bb81e29323e40b4de66f331
Closes-Bug: 1361448