Updated openstack/openstack Project: openstack/keystone  62a1672aa2f0cbab46a2df886f278487d94c24c9
Remove assignment_api dependency on token_api

The assignment_api now uses the callback system to avoid needing
to directly call token_api for invalidating the tokens in the
token persistence system.

This change is in support of non-persistent tokens which has a
dependency on unifying interactions with the token persistence
system.

Change-Id: I66af0a7281cf9a4d451f0f6d86da8182e2cc0e60
bp: non-persistent-tokens