Updated openstack/openstack Project: openstack/keystone  18efc7827a2b6767182772134bd00d1c110de7f7
Mark kvs backends as deprecated, for removal in Kilo.

Although kvs.Base was marked from removal in Juno, the backends
themselves were not explicitly marked. This patch does this,
as well as extendes kvs.Base also to Kilo so that they can all
be removed together.

Partially implements: bp deprecated-as-of-juno
Change-Id: Ia2e7c95074c4b6b9743c6086842105a5579fc0c9