Updated openstack/openstack Project: openstack/neutron  2faa4a4701bd9fc7b63c8b1b319eb5659bf32ef0
Fix l3 agent scheduling logic to avoid unwanted failures

In case router is being added to l3 agent which is already hosting
the router it is fine to let such a request to succeed.
This patch also adds a check for unnecessary scheduling that might happen
twice in described case and lead to unwanted messages in the logs.

Change-Id: Id104b36ba7e1e6f6a9378ee600c33e9962230521
Closes-Bug: #1358636